Add honeycomb service
