+++ logback-honeycomb-client/src/main/java/com/tersesystems/logback/honeycomb/client/HoneycombClientService.java
@@ -1,0 +1,5 @@
+package com.tersesystems.logback.honeycomb.client;
+
+public interface HoneycombClientService {
+    HoneycombClient newClient();
+}
