+++ logback-honeycomb-okhttp/src/main/java/com/tersesystems/logback/honeycomb/okhttp/HoneycombOkHTTPClientService.java
@@ -1,0 +1,13 @@
+package com.tersesystems.logback.honeycomb.okhttp;
+
+import com.fasterxml.jackson.core.JsonFactory;
+import com.tersesystems.logback.honeycomb.client.HoneycombClient;
+import com.tersesystems.logback.honeycomb.client.HoneycombClientService;
+import okhttp3.OkHttpClient;
+
+public class HoneycombOkHTTPClientService implements HoneycombClientService {
+    @Override
+    public HoneycombClient newClient() {
+        return new HoneycombOkHTTPClient(new OkHttpClient(), new JsonFactory());
+    }
+}
