+++ logback-honeycomb-okhttp/src/main/resources/META-INF/services/com.tersesystems.logback.honeycomb.client.HoneycombClientService
@@ -1,0 +1,1 @@
+com.tersesystems.logback.honeycomb.okhttp.HoneycombOkHTTPClientService
