+++ logback-honeycomb-playws_2.12/logback-honeycomb-playws_2.12.gradle
@@ -26,9 +26,6 @@ dependencies {
     // https://mvnrepository.com/artifact/com.typesafe.play/play-ws-standalone
     implementation group: 'com.typesafe.play', name: 'play-ahc-ws-standalone_2.12', version: '2.0.7'
 
-    compile group: 'com.google.auto.service', name: 'auto-service', version: '1.0-rc6'
-    annotationProcessor "com.google.auto.service:auto-service:1.0-rc6"
-
     testImplementation project(":logback-tracing")
     testImplementation project(":logback-classic")
     testImplementation project(":logback-honeycomb-appender")
