+++ logback-honeycomb-playws_2.12/src/main/java/com/tersesystems/logback/honeycomb/playws/HoneycombPlayWSClient.java
@@ -11,8 +11,6 @@
 package com.tersesystems.logback.honeycomb.playws;
 
 import akka.actor.ActorSystem;
-import akka.stream.ActorMaterializer;
-import akka.stream.ActorMaterializerSettings;
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonParser;
@@ -22,50 +20,27 @@ import com.tersesystems.logback.honeycomb.client.HoneycombClient;
 import com.tersesystems.logback.honeycomb.client.HoneycombHeaders;
 import com.tersesystems.logback.honeycomb.client.HoneycombRequest;
 import com.tersesystems.logback.honeycomb.client.HoneycombResponse;
-import com.typesafe.config.Config;
-import com.typesafe.config.ConfigFactory;
+import play.libs.ws.*;
+
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.time.Instant;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.CompletionStage;
 import java.util.function.Function;
-import play.api.libs.ws.ahc.AhcWSClientConfig;
-import play.api.libs.ws.ahc.AhcWSClientConfigFactory;
-import play.libs.ws.DefaultBodyReadables;
-import play.libs.ws.DefaultBodyWritables;
-import play.libs.ws.StandaloneWSResponse;
-import play.libs.ws.ahc.StandaloneAhcWSClient;
-import play.libs.ws.ahc.StandaloneAhcWSRequest;
 
-@AutoService(HoneycombClient.class)
 public class HoneycombPlayWSClient
     implements HoneycombClient, DefaultBodyWritables, DefaultBodyReadables {
 
-  public static final String DEFAULT_ACTORSYSTEM_NAME = "honeycombClientSystem";
-  private static final String AKKA_MAX_THREADS_KEY =
-      "akka.actor.default-dispatcher.fork-join-executor.parallelism-max";
-
-  private final StandaloneAhcWSClient client;
+  private final StandaloneWSClient client;
   private final JsonFactory factory = new JsonFactory();
   private final ActorSystem actorSystem;
+  private final boolean terminateOnClose;
 
-  public HoneycombPlayWSClient() {
-    Map<String, Object> clientMap = new HashMap<>();
-    clientMap.put("play.ws.compressionEnabled", Boolean.TRUE);
-    clientMap.put("play.ws.useragent", "Logback Honeycomb Client");
-    clientMap.put(AKKA_MAX_THREADS_KEY, 2);
-    Config config = ConfigFactory.parseMap(clientMap).withFallback(ConfigFactory.load());
-
-    actorSystem = ActorSystem.create(DEFAULT_ACTORSYSTEM_NAME, config);
-    AhcWSClientConfig ahcWsClientConfig =
-        AhcWSClientConfigFactory.forConfig(config, config.getClass().getClassLoader());
-    this.client = StandaloneAhcWSClient.create(ahcWsClientConfig, createMaterializer(actorSystem));
+  public HoneycombPlayWSClient(StandaloneWSClient client, ActorSystem actorSystem, boolean terminateOnClose) {
+    this.client = client;
+    this.actorSystem = actorSystem;
+    this.terminateOnClose = terminateOnClose;
   }
 
   /** Posts a single event to honeycomb, using the "1/events" endpoint. */
@@ -76,11 +51,11 @@ public class HoneycombPlayWSClient
       HoneycombRequest<E> request,
       Function<HoneycombRequest<E>, byte[]> encodeFunction) {
     String honeycombURL = eventURL(dataset);
-    StandaloneAhcWSRequest ahcWSRequest = client.url(honeycombURL);
+    StandaloneWSRequest wsRequest = client.url(honeycombURL);
     byte[] bytes = encodeFunction.apply(request);
-    return ahcWSRequest
+    return wsRequest
         .addHeader(HoneycombHeaders.teamHeader(), apiKey)
-        .addHeader(HoneycombHeaders.eventTimeHeader(), isoTime(request.getTimestamp()))
+        .addHeader(HoneycombHeaders.eventTimeHeader(), HoneycombRequestFormatter.isoTime(request.getTimestamp()))
         .addHeader(HoneycombHeaders.sampleRateHeader(), request.getSampleRate().toString())
         .addHeader("Content-Type", "application/json")
         .post(body(bytes))
@@ -95,9 +70,9 @@ public class HoneycombPlayWSClient
       Function<HoneycombRequest<E>, byte[]> encodeFunction) {
     String honeycombURL = batchURL(dataset);
     try {
-      StandaloneAhcWSRequest ahcWSRequest = client.url(honeycombURL);
+      StandaloneWSRequest wsRequest = client.url(honeycombURL);
       byte[] batchedJson = generateBatchJson(requests, encodeFunction);
-      return ahcWSRequest
+      return wsRequest
           .addHeader(HoneycombHeaders.teamHeader(), apiKey)
           .addHeader("Content-Type", "application/json")
           .post(body(batchedJson))
@@ -131,7 +106,7 @@ public class HoneycombPlayWSClient
         }
       }
     } catch (IOException e) {
-
+      throw new IllegalStateException("should never happen", e);
     }
 
     return list;
@@ -147,14 +122,11 @@ public class HoneycombPlayWSClient
     return batchURL + dataset;
   }
 
-  private ActorMaterializer createMaterializer(ActorSystem system) {
-    return ActorMaterializer.create(
-        ActorMaterializerSettings.create(system), system, system.name());
-  }
-
   public void close() throws IOException {
     client.close();
-    actorSystem.terminate();
+    if (terminateOnClose) {
+      actorSystem.terminate();
+    }
   }
 
   private <E> byte[] generateBatchJson(
@@ -174,38 +146,4 @@ public class HoneycombPlayWSClient
     return stream.toByteArray();
   }
 
-  class HoneycombRequestFormatter<E> {
-    private final JsonGenerator generator;
-    private final Function<HoneycombRequest<E>, byte[]> encodeFunction;
-
-    HoneycombRequestFormatter(
-        JsonGenerator generator, Function<HoneycombRequest<E>, byte[]> encodeFunction) {
-      this.generator = generator;
-      this.encodeFunction = encodeFunction;
-    }
-
-    void start() throws IOException {
-      this.generator.writeStartArray();
-    }
-
-    void end() throws IOException {
-      this.generator.writeEndArray();
-    }
-
-    void format(HoneycombRequest request) throws IOException {
-      byte[] bytes = encodeFunction.apply(request);
-
-      generator.writeStartObject();
-      generator.writeStringField("time", isoTime(request.getTimestamp()));
-      generator.writeNumberField("samplerate", request.getSampleRate());
-      generator.writeFieldName("data");
-      generator.writeRaw(":");
-      generator.writeRaw(new String(bytes, StandardCharsets.UTF_8));
-      generator.writeEndObject();
-    }
-  }
-
-  private String isoTime(Instant eventTime) {
-    return DateTimeFormatter.ISO_INSTANT.format(eventTime);
-  }
 }
