+++ logback-honeycomb-playws_2.12/src/main/resources/META-INF/services/com.tersesystems.logback.honeycomb.client.HoneycombClientService
@@ -1,0 +1,1 @@
+com.tersesystems.logback.honeycomb.playws.HoneycombPlayWSClientService
